Consolidates WTR configuration
